sprint 5-6

+++ internal/model/Order.go
@@ -12,7 +12,7 @@ const (
 type DataOrder struct {
 	Number     string    `json:"number"`
 	Status     string    `json:"status"`
-	Accrual    *int      `json:"accrual,omitempty"` //nullable
+	Accrual    *float64  `json:"accrual,omitempty"` //nullable
 	Person     string    `json:"-"`
 	UploadedAt time.Time `json:"uploaded_at"`
 }
